III-3141 Remove the abstract OfferHistoryProjector
It was only implemented by one child class
